ref(lesson-content): changes position of Self-regulated Question Text field.
