+++ libs/feature/teaching/src/lib/lesson/forms/lesson-content.tsx
@@ -6,7 +6,7 @@ import {
 	LessonContentType,
 	ValueByContentType
 } from "@self-learning/types";
-import { RemovableTab, SectionHeader, Tabs } from "@self-learning/ui/common";
+import { Divider, RemovableTab, SectionHeader, Tabs } from "@self-learning/ui/common";
 import { Reorder } from "framer-motion";
 import { useCallback, useEffect, useMemo, useState } from "react";
 import { Control, Controller, useFieldArray, useFormContext } from "react-hook-form";
@@ -15,7 +15,8 @@ import { IFrameInput } from "../content-types/iframe";
 import { PdfInput } from "../content-types/pdf";
 import { VideoInput } from "../content-types/video";
 import { LessonFormModel } from "@self-learning/teaching";
-import { Form, MarkdownField } from "@self-learning/ui/forms";
+import { Form, LabeledField, MarkdownField } from "@self-learning/ui/forms";
+import { Do_Hyeon } from "next/dist/compiled/@next/font/dist/google";
 
 export type SetValueFn = <CType extends LessonContentType["type"]>(
 	type: CType,
@@ -236,7 +237,10 @@ function RenderContentType({ index, content }: { index: number; content: LessonC
 }
 
 function LessonDescriptionForm() {
-	const { control } = useFormContext<LessonFormModel>();
+	const form = useFormContext<LessonFormModel>();
+	const control = form.control;
+
+	const currentLessonType = form.watch("lessonType");
 
 	return (
 		<section>
@@ -251,7 +255,25 @@ function LessonDescriptionForm() {
 					render={({ field }) => (
 						<MarkdownField content={field.value as string} setValue={field.onChange} />
 					)}
-				></Controller>
+				/>
+
+				<div className={"py-4"}>
+					{currentLessonType === "SELF_REGULATED" && (
+						<LabeledField label="Aktivierungsfrage" optional={false}>
+							<Controller
+								control={control}
+								name="selfRegulatedQuestion"
+								render={({ field }) => (
+									<MarkdownField
+										content={field.value as string}
+										setValue={field.onChange}
+										inline={true}
+									/>
+								)}
+							/>
+						</LabeledField>
+					)}
+				</div>
 			</Form.MarkdownWithPreviewContainer>
 		</section>
 	);
